+++ ec2_metadata.py
@@ -19,12 +19,15 @@ METADATA_URL = SERVICE_URL + 'meta-data/'
 USERDATA_URL = SERVICE_URL + 'user-data/'
 
 
-class EC2Metadata(object):
+class BaseLazyObject(object):
 
     def clear_all(self):
         for key in tuple(self.__dict__.keys()):
             if isinstance(getattr(self.__class__, key), cached_property):
                 del self.__dict__[key]
+
+
+class EC2Metadata(BaseLazyObject):
 
     @property
     def account_id(self):
@@ -63,26 +66,10 @@ class EC2Metadata(object):
         return requests.get(METADATA_URL + 'mac').text
 
     @cached_property
-    def network(self):
-        '''The network interface map is keyed by MAC address. This structure will
-           be returned like how it is given in the metadata interface.'''
-        return self.recursive_get_data(METADATA_URL, 'network/interfaces/macs/')
-
-    # helper function, not really designed to be called directly by clients.
-    # this works by recognizing that value-only keys don't have a trailing slash,
-    # and keys with a trailing slash have subkeys.
-    def recursive_get_data(self, baseurl, key):
-        curr_url = baseurl + key
-        if not key.endswith('/'):  # don't explore further, just fetch
-            return requests.get(curr_url).text
-
-        # trailing slash, need to explore further
-        ret = {}
-        for l in requests.get(curr_url).text.splitlines():
-            lkey = l.rstrip('/')
-            ret[lkey] = self.recursive_get_data(curr_url, l)
-
-        return ret
+    def network_interfaces(self):
+        macs_text = requests.get(METADATA_URL + 'network/interfaces/macs/').text
+        macs = [line.rstrip('/') for line in macs_text.splitlines()]
+        return {mac: NetworkInterface(mac) for mac in macs}
 
     @cached_property
     def private_hostname(self):
@@ -119,6 +106,117 @@ class EC2Metadata(object):
             return None
         else:
             return resp.content
+
+
+class NetworkInterface(BaseLazyObject):
+
+    def __init__(self, mac):
+        self.mac = mac
+
+    def __repr__(self):
+        return 'NetworkInterface({mac})'.format(mac=repr(self.mac))
+
+    def __eq__(self, other):
+        return isinstance(other, NetworkInterface) and self.mac == other.mac
+
+    def _url(self, item):
+        return '{base}network/interfaces/macs/{mac}/{item}'.format(
+            base=METADATA_URL,
+            mac=self.mac,
+            item=item
+        )
+
+    @cached_property
+    def device_number(self):
+        return int(requests.get(self._url('device-number')).text)
+
+    @cached_property
+    def ipv4_associations(self):
+        associations = {}
+        for public_ip in self.public_ipv4s:
+            resp = requests.get(self._url('ipv4-associations/{}'.format(public_ip)))
+            resp.raise_for_status()
+            private_ips = resp.text.splitlines()
+            associations[public_ip] = private_ips
+        return associations
+
+    # No IPV6 instances at hand to test this on, so I only know you get 404 in
+    # case there are none
+    # @cached_property
+    # def ipv6s(self):
+    #     pass
+
+    @cached_property
+    def owner_id(self):
+        return requests.get(self._url('owner-id')).text
+
+    @cached_property
+    def private_hostname(self):
+        return requests.get(self._url('local-hostname')).text
+
+    @cached_property
+    def private_ipv4s(self):
+        return requests.get(self._url('local-ipv4s')).text.splitlines()
+
+    @cached_property
+    def public_hostname(self):
+        return requests.get(self._url('public-hostname')).text
+
+    @cached_property
+    def public_ipv4s(self):
+        return requests.get(self._url('public-ipv4s')).text.splitlines()
+
+    @cached_property
+    def security_groups(self):
+        return requests.get(self._url('security-groups')).text.splitlines()
+
+    @cached_property
+    def security_group_ids(self):
+        return requests.get(self._url('security-group-ids')).text.splitlines()
+
+    @cached_property
+    def subnet_id(self):
+        return requests.get(self._url('subnet-id')).text
+
+    @cached_property
+    def subnet_ipv4_cidr_block(self):
+        resp = requests.get(self._url('subnet-ipv4-cidr-block'))
+        if resp.status_code == 404:
+            return None
+        else:
+            return resp.text
+
+    # No IPV6 instances at hand to test this on, so I only know you get 404 in
+    # case there are none
+    # @cached_property
+    # def subnet_ipv6_cidr_blocks(self):
+    #     pass
+
+    @cached_property
+    def vpc_id(self):
+        return requests.get(self._url('vpc-id')).text
+
+    @cached_property
+    def vpc_ipv4_cidr_block(self):
+        resp = requests.get(self._url('vpc-ipv4-cidr-block'))
+        if resp.status_code == 404:
+            return None
+        else:
+            return resp.text
+
+    @cached_property
+    def vpc_ipv4_cidr_blocks(self):
+        resp = requests.get(self._url('vpc-ipv4-cidr-blocks'))
+        if resp.status_code == 404:
+            return []
+        else:
+            return resp.text.splitlines()
+
+    # No IPV6 at hand to test this on, so I only know you get 404 in case there
+    # are none
+    # @cached_property
+    # def vpc_ipv6_cidr_blocks(self):
+    #     pass
 
 
 ec2_metadata = EC2Metadata()