+++ HISTORY.rst
@@ -8,7 +8,8 @@ Pending Release
 
 .. Insert new release notes below this line
 
-* Add ``network``` attribute.
+* Add ``network_interfaces`` attribute which is a list of ``NetworkInterface``
+  instances, which have many attributes themselves.
 
 1.1.0 (2017-08-07)
 ------------------